feature: added Error_Fragment and its layout to show when something bad goes on when calling the nasa api.
SharedPrefs: added sharedPrefs to stop calling the nasa api everytime it opens the app, if the user already opened the app, then it should save the picture of the day inside sharedPrefs to restore it when the user change tabs or close the app.

added RetryCallBack.kt to ErrorFragment.kt.